+++ park/src/main/java/com/park/park/entities/KolejkiGorskieEntity.java
@@ -5,7 +5,6 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
-import java.math.BigInteger;
 import java.sql.Date;
 
 @Entity
@@ -15,30 +14,30 @@ import java.sql.Date;
 @Table(name = "KOLEJKI_GORSKIE", schema = "PARK_ROZRYWKI")
 public class KolejkiGorskieEntity {
     @Id
-    @Column(name = "ID_ATRAKCJI", nullable = false, precision = 0)
-    private BigInteger idAtrakcji;
+    @Column(name = "ID_ATRAKCJI", nullable = false)
+    private long idAtrakcji;
     @Basic
-    @Column(name = "CZAS_OCZEKIWANIA", nullable = false, precision = 0)
-    private BigInteger czasOczekiwania;
+    @Column(name = "CZAS_OCZEKIWANIA", nullable = false)
+    private int czasOczekiwania;
     @Basic
-    @Column(name = "MAKSYMALNA_WYSOKOSC", nullable = false, precision = 2)
-    private short maksymalnaWysokosc;
+    @Column(name = "MAKSYMALNA_WYSOKOSC", nullable = false)
+    private int maksymalnaWysokosc;
     @Basic
-    @Column(name = "MINIMALNY_WZROST", nullable = false, precision = 2)
-    private short minimalnyWzrost;
+    @Column(name = "MINIMALNY_WZROST", nullable = false)
+    private int minimalnyWzrost;
     @Basic
     @Column(name = "SREDNIA_PREDKOSC", nullable = false, precision = 2)
-    private short sredniaPredkosc;
+    private float sredniaPredkosc;
     @Basic
     @Column(name = "CZAS_JAZDY", nullable = false, precision = 2)
-    private short czasJazdy;
+    private float czasJazdy;
     @Basic
-    @Column(name = "POJEMNOSC_WAGONIKA", nullable = false, precision = 0)
-    private BigInteger pojemnoscWagonika;
+    @Column(name = "POJEMNOSC_WAGONIKA", nullable = false)
+    private int pojemnoscWagonika;
     @Basic
     @Column(name = "DATA_OSTATNIEGO_PRZEGLADU", nullable = false)
     private Date dataOstatniegoPrzegladu;
     @Basic
-    @Column(name = "MINIMIALNY_WIEK", nullable = false, precision = 0)
-    private BigInteger minimialnyWiek;
+    @Column(name = "MINIMIALNY_WIEK", nullable = false)
+    private int minimialnyWiek;
 }
